refactor: migrate LoginV2 to Login
