--- src/domains/users/Auth/components/Login/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Login";
