--- src/pages/loginV2.tsx
@@ -1,7 +1,0 @@
-import { LoginFormV2 } from "@Users/Auth/components";
-
-const Login = () => {
-  return <LoginFormV2 />;
-};
-
-export default Login;
